requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
langchain-community
pypdf
python-dotenv
pydantic
python-multipart
